Removed the AmazonPersistenceModule
Not required by Spring
